minor Codebase cleanup
Corrected some project file issues and removed some files no longer
needed and organized some files into folders.

+++ src/Game.h
@@ -143,7 +143,7 @@ public:
 private:
 	Gamestate _State;
 	World _World;
-    int _Progress;
+    double _Progress;
 
 };
